Refactor: moved IOdometerRepository setup code to a fixture class
